Fix issue #17507017: took ~3 seconds to move an app to the foreground
This is because of the 5 second timeout from when the user can bring
home to the foreground until regular third party apps can launch an
activity on top of it.  Activities launched from notifications look
like they are being launched by the app, so get impacted by the timeout.

Fix this be also looking at the actual caller to see if they are
allowed to pop in front regardless of the timeout.

Change-Id: I63fbc2bcabf585e6d2810a2309f0613fdf91fdf5