Fixed issue with moveTaskToBack on single stack devices
Icdad980eec64e081d15679600da07cf4431e40d6 allowed us to
properly return to the home acitvity when a task is moved
to back. However, this improperly moved the home task to
the front if it is the task we are moving to the back on
a single stack device. We now prevent the movement of the
home task to the front from happening.

Bug: 23088310
Change-Id: Ic21779cdb2d2007671d212d41fab5e68be2ae632