Don't move task if it is already in the destination stack
When moving a task from one stack to another we detach the source
stack from the display if the stack is empty after the task move.
If the source and destination stack are the same we detach the
stack from the display when the task is removed, but it isn't
re-attached to the display when the task is add back since we
only attach stacks to displays when they are created.
We no longer try to move the task if it is already in the destination
stack.

Bug: 25671469
Change-Id: Idde54d362a35716223bc5f1143000111af380239