Ensure visibility of pinned activity when starting it.
Currently we depend on resuming focused activity within the focused
stack to make it visible. This doesn't work for pinned stack, because
neither the stack nor the activity are focused and as a result the
pinned activity doesn't get resumed and made visible. We need to
explicitly force visibility without focusing the stack.

This fixes two issues:
* pinned activity not showing up when launched from a different full
screen stack activity;
* no entry animation when launching the activity from another full
screen stack.

Bug: 26891834
Bug: 25672053
Change-Id: I83065e10a2d237f1a1f26115e94c6648570e1d7f