Do remove starting windows for activities in an invisible stack
Depending on app launch timing it is possible for the initializing
activity that is the top of an invisible stack to still have its
starting window visible since we skip removing orphaned starting
window for the top activity in a stack. We now remove starting
windows for all activities in an invisible stack.

Bug: 28173906
Change-Id: Ie5284c9681d06969f647cbe46a77dd072ca61435