Allow no longer waiting for visible activity to be stopped.
When process the stop list we remove an activity that was previously
waiting for visible and is now visible from the list and also hide
its windows if it is finishing. However, we don't transition it to the
stopped state because it was waiting for visible. If it is finishing
and the windows are hidden then it makes sense to transition it into
the stopped state since it will not be visible to the user again.

Bug: 30058392
Change-Id: Iac81e2877716ae58d05aed42e1433190a2f70428